+++ gateway/src/main.rs
@@ -2,7 +2,8 @@ extern crate gateway;
 
 use crossbeam_channel::{unbounded, Receiver, Sender};
 
-use gateway::model::{KafkaCommands, KafkaEvents};
+use gateway::kafka_events::KafkaEvents;
+use gateway::model::KafkaCommands;
 use gateway::router::RouterCommand;
 use gateway::websocket::WsSession;
 use gateway::{kafka_io, router};
